% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_table_.R
\name{scenario_table_Input}
\alias{scenario_table_Input}
\title{Scenario Table Input}
\usage{
scenario_table_Input(
  inputId,
  data_,
  individual_id_options,
  population_id_options
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
